Add Dtos
Some of the processing of records on the client side was getting a little clunky, so I added DTOs to make the Client-side code a little easier.

I skipped adding "DTO" to all of the class names because I wanted to avoid redundant naming (ex. Dto.TitleDto)

Although this makes it a little cumbersome when they appear where the API.Models are, it should not be confusing as long as you specificy the namespace. May revisit this.