So I ran npm eject...
This had the consequence of shoving the Webpack and other configuration files into my source.
I did this because I really wanted to use Less, but Material UI actually has a more handy styling system.

I got a lot of warnings that npm Eject was non "undoable", so I decided to just leave it since it was working fine (last I checked).