Fleshed out the Title View
This commit is basically just "all of the work" for the UI side of things, including looking up Titles and viewing their details.

I kind of wanted to keep everything on one page to keep it simple, but I underestimated how much metadata exists with each title.

I will revisit this.